requirements.txt
altgraph==0.10.2
awscli==1.10.65
bcrypt==3.1.1
bdist-mpkg==0.5.0
boto==2.44.0
botocore==1.4.55
cffi==1.8.3
colorama==0.3.7
cryptography==1.5.2
Django==1.11.5
django-debug-toolbar==1.9.1
django-simple-history==2.3.0
docutils==0.12
enum34==1.1.6
futures==3.0.5
httplib2==0.9.2
idna==2.1
ipaddress==1.0.17
jmespath==0.9.0
macholib==1.5.1
#matplotlib==1.3.1
modulegraph==0.10.4
mysql==0.0.1
MySQL-python==1.2.5
numpy==1.8.0
paramiko==2.0.8
py==1.4.32
py2app==0.7.3
pyasn1==0.1.9
pycparser==2.17
PyJWT==1.4.2
pyOpenSSL==0.14
#pyobjc-core==2.5.1
#pyobjc-framework-Accounts==2.5.1
#pyobjc-framework-AddressBook==2.5.1
#pyobjc-framework-AppleScriptKit==2.5.1
#pyobjc-framework-AppleScriptObjC==2.5.1
#pyobjc-framework-Automator==2.5.1
#pyobjc-framework-CFNetwork==2.5.1
#pyobjc-framework-Cocoa==2.5.1
#pyobjc-framework-Collaboration==2.5.1
#pyobjc-framework-CoreData==2.5.1
#pyobjc-framework-CoreLocation==2.5.1
#pyobjc-framework-CoreText==2.5.1
#pyobjc-framework-DictionaryServices==2.5.1
#pyobjc-framework-EventKit==2.5.1
#pyobjc-framework-ExceptionHandling==2.5.1
#pyobjc-framework-FSEvents==2.5.1
#pyobjc-framework-InputMethodKit==2.5.1
#pyobjc-framework-InstallerPlugins==2.5.1
#pyobjc-framework-InstantMessage==2.5.1
#pyobjc-framework-LatentSemanticMapping==2.5.1
#pyobjc-framework-LaunchServices==2.5.1
#pyobjc-framework-Message==2.5.1
#pyobjc-framework-OpenDirectory==2.5.1
#pyobjc-framework-PreferencePanes==2.5.1
#pyobjc-framework-PubSub==2.5.1
#pyobjc-framework-QTKit==2.5.1
#pyobjc-framework-Quartz==2.5.1
#pyobjc-framework-ScreenSaver==2.5.1
#pyobjc-framework-ScriptingBridge==2.5.1
#pyobjc-framework-SearchKit==2.5.1
#pyobjc-framework-ServiceManagement==2.5.1
#pyobjc-framework-Social==2.5.1
#pyobjc-framework-SyncServices==2.5.1
#pyobjc-framework-SystemConfiguration==2.5.1
#pyobjc-framework-WebKit==2.5.1
pyOpenSSL==0.14
pyparsing==2.0.1
pysftp==0.2.9
pytest==3.0.6
python-dateutil==2.6.0
python-openid==2.2.5
pytz==2013.7
requests==2.11.1
requests-oauthlib==0.7.0
rsa==3.4.2
s3transfer==0.1.4
scipy==0.13.0
six==1.10.0
SQLAlchemy==1.1.3
twilio==5.6.0
xattr==0.6.4
zope.interface==4.1.1
django-allauth==0.30.0
django-bootstrap-form==3.2.1
git+https://github.com/bandwidth/python-bandwidth.git#egg=bandwidth_sdk
